update volume to be double

+++ Models/Crafts/ProfitableCrafts.cs
@@ -10,7 +10,7 @@ namespace Coflnet.Sky.Crafts.Models
         public IEnumerable<Ingredient> Ingredients { get; set; }
         public RequiredCollection ReqCollection { get; set; }
         public RequiredCollection ReqSlayer { get; set; }
-        public float Volume { get; set; }
+        public double Volume { get; set; }
         public float Median { get; set; }
     }
 }
